feat(remove-destiny-routes): Remove legacy Destiny routes. Only use the Destiny database for Grimoire Cards.
