+++ destiny/destiny.controller.js
@@ -46,7 +46,6 @@ class DestinyController {
      * Get characters for the current user.
      *
      * @returns {*|Array}
-     * @private
      */
     async getCharacters(req, res) {
         const { session: { displayName, membershipType } } = req;
@@ -77,94 +76,6 @@ class DestinyController {
         });
 
         res.json(characterBases);
-    }
-
-    /**
-     * Get the currently available field test weapons from the gun smith.
-     *
-     * @param req
-     * @param res
-     */
-    async getFieldTestWeapons(req, res) {
-        const { session: { displayName, membershipType } } = req;
-        const { bungie: { access_token: accessToken }, membershipId } = await this
-            .users.getUserByDisplayName(displayName, membershipType);
-        const characters = await this.destiny.getCharacters(membershipId, membershipType);
-
-        if (characters && characters.length > 0) {
-            const { characterBase: { characterId } } = characters[0];
-            const vendor = await this
-                .destiny.getFieldTestWeapons(characterId, membershipType, accessToken);
-            const { itemHashes } = vendor;
-            const items = await Promise.all(
-                itemHashes.map(itemHash => this.world.getItemByHash(itemHash)),
-            );
-
-            res.json(items.map(item => item.itemName));
-        }
-
-        return res.status(411).end();
-    }
-
-    /**
-     * Get the currently available foundry orders from the gun smith.
-     *
-     * @param req
-     * @param res
-     */
-    async getFoundryOrders(req, res) {
-        const { session: { displayName, membershipType } } = req;
-        const { bungie: { access_token: accessToken }, membershipId } = await this
-            .users.getUserByDisplayName(displayName, membershipType);
-        const characters = await this.destiny.getCharacters(membershipId, membershipType);
-
-        if (characters && characters.length > 0) {
-            const { characterBase: { characterId } } = characters[0];
-            const vendor = await this.destiny.getFoundryOrders(characterId, accessToken);
-            const { itemHashes } = vendor;
-            const items = await Promise.all(itemHashes.map(itemHash => this
-                .world.getItemByHash(itemHash)));
-
-            res.json(items.map(item => item.itemName));
-        }
-
-        return res.status(411).end();
-    }
-
-    /**
-     * Get the currently available iron banner rewards from Lord Saladin.
-     *
-     * @param req
-     * @param res
-     */
-    async getIronBannerEventRewards(req, res) {
-        const { session: { displayName, membershipType } } = req;
-        const { bungie: { access_token: accessToken }, membershipId } = await this
-            .users.getUserByDisplayName(displayName, membershipType);
-        const characters = await this.destiny.getCharacters(membershipId, membershipType);
-        const characterItems = await Promise.all(characters
-            .map(({ characterBase: { characterId } }) => this
-                .destiny.getIronBannerEventRewards(characterId, membershipType, accessToken)));
-        const rewards = _.flatten(characterItems);
-        const itemHashes = _.uniq(rewards.map(({ item = {} }) => item.itemHash));
-        const items = await Promise.all(itemHashes.map(itemHash => this
-            .world.getItemByHash(itemHash)));
-        const weapons = _.filter(items, item => _.contains(item.itemCategoryHashes, 1));
-        const hunterArmor = _.filter(items, item => _.contains(item.itemCategoryHashes, 20)
-            && _.contains(item.itemCategoryHashes, 23));
-        const titanArmor = _.filter(items, item => _.contains(item.itemCategoryHashes, 20)
-            && _.contains(item.itemCategoryHashes, 22));
-        const warlockArmor = _.filter(items, item => _.contains(item.itemCategoryHashes, 20)
-            && _.contains(item.itemCategoryHashes, 21));
-
-        res.json({
-            weapons: _.map(weapons, item => item.itemName),
-            armor: {
-                hunter: _.map(hunterArmor, item => item.itemName),
-                titan: _.map(titanArmor, item => item.itemName),
-                warlock: _.map(warlockArmor, item => item.itemName),
-            },
-        });
     }
 
     /**
@@ -199,40 +110,11 @@ class DestinyController {
     }
 
     /**
-     * Get the exotic weapons and gear available from Xur.
-     * @param req
-     * @param res
-     */
-    async getXur(req, res) {
-        const vendor = await this.destiny.getXur();
-        const { itemHashes, nextRefreshDate } = vendor;
-
-        if (itemHashes === undefined || itemHashes.length === 0) {
-            return res.status(200).json({ itemHashes: [], nextRefreshDate });
-        }
-
-        const allItems = await Promise.all(itemHashes.map(itemHash => this
-            .world.getItemByHash(itemHash)));
-        const itemPromises = _.map(allItems, item => {
-            if (item.itemName === 'Exotic Engram'
-                || item.itemName === 'Legacy Engram') {
-                return this.world.getItemByHash(item.itemHash)
-                    .then(itemDetail => item.itemName.replace('Engram', '')
-                            + itemDetail.itemTypeName);
-            }
-
-            return Promise.resolve(item.itemName);
-        });
-        const items = await Promise.all(itemPromises);
-
-        return res.json(items);
-    }
-
-    /**
-     * Insert or update the Destiny manifest.
+     * Insert or update the Destiny manifest if needed.
      */
     async upsertManifest(req, res) {
         const manifest = await this.destiny.getManifest(true);
+
         await this.world.updateManifest(manifest);
 
         res.status(200).json(manifest);
