+++ helpers/world2.spec.js
@@ -2,6 +2,7 @@
  * World Model Tests
  */
 const World = require('./world2');
+const { xurHash } = require('../destiny2/destiny2.constants');
 
 const world = new World({
     directory: process.env.DESTINY2_DATABASE_DIR,
@@ -27,6 +28,43 @@ describe('It\'s Bungie\'s 2nd world. You\'re just querying it.', () => {
                 const { displayProperties: { name } } = category;
 
                 expect(name).toEqual('Hand Cannon');
+                done();
+            })
+            .catch(err => {
+                done(err);
+            });
+    });
+
+    it('should return the Hunter character class', done => {
+        world.getClassByHash(671679327)
+            .then(characterClass => {
+                const { displayProperties: { name } } = characterClass;
+
+                expect(name).toEqual('Hunter');
+                done();
+            })
+            .catch(err => {
+                done(err);
+            });
+    });
+
+    it('should return Night Watch', done => {
+        const itemName = 'Night Watch';
+
+        world.getItemByName(itemName)
+            .then(items => {
+                expect(items[0].displayProperties.name).toEqual(itemName);
+                done();
+            })
+            .catch(err => {
+                done(err);
+            });
+    });
+
+    it('should return the icon of the Agent of Nine', done => {
+        world.getVendorIcon(xurHash)
+            .then(url => {
+                expect(url).toBeDefined();
                 done();
             })
             .catch(err => {
